refactor: move `JWT_SECRET` to environment variables and remove `env.ts`
