--- packages/server/src/env.ts
@@ -1,1 +1,0 @@
-export const JWT_SECRET = 'this-is-sample-secret-key';
